+++ src/sources/soundsourceffmpeg.cpp
@@ -53,10 +53,13 @@ constexpr int64_t kavStreamDecoderFrameDelayAAC = 2112;
 
 constexpr SINT kMaxSamplesPerMP3Frame = 1152;
 
+// Note: The fist audio stream can be at any stream index after other stream types
+constexpr int kFirstAudioStream = -1;
+
 const Logger kLogger("SoundSourceFFmpeg");
 
 int64_t getStreamStartTime(const AVStream& avStream) {
-    auto start_time = avStream.start_time;
+    int64_t start_time = avStream.start_time;
     if (start_time == AV_NOPTS_VALUE) {
         // This case is not unlikely, e.g. happens when decoding WAV files.
         switch (avStream.codecpar->codec_id) {
@@ -154,12 +157,14 @@ void SoundSourceFFmpeg::initChannelLayoutFromStream(
         // layout, e.g. for a mono WAV files with a single channel!
         av_channel_layout_default(pUninitializedChannelLayout,
                 avStream.codecpar->ch_layout.nb_channels);
-        kLogger.info()
-                << "Unknown channel layout -> using default layout"
-                << pUninitializedChannelLayout->order
-                << "for"
-                << avStream.codecpar->ch_layout.nb_channels
-                << "channel(s)";
+        if (avStream.codecpar->ch_layout.nb_channels > 1) {
+            kLogger.warning()
+                    << "Unknown channel layout -> using default layout"
+                    << pUninitializedChannelLayout->order
+                    << "for"
+                    << avStream.codecpar->ch_layout.nb_channels
+                    << "channels";
+        }
     } else {
         av_channel_layout_default(pUninitializedChannelLayout, 0);
         av_channel_layout_copy(pUninitializedChannelLayout, &avStream.codecpar->ch_layout);
@@ -173,12 +178,14 @@ int64_t SoundSourceFFmpeg::getStreamChannelLayout(const AVStream& avStream) {
         // Workaround: FFmpeg sometimes fails to determine the channel
         // layout, e.g. for a mono WAV files with a single channel!
         channel_layout = av_get_default_channel_layout(avStream.codecpar->channels);
-        kLogger.info()
-                << "Unknown channel layout -> using default layout"
-                << channel_layout
-                << "for"
-                << avStream.codecpar->channels
-                << "channel(s)";
+        if (avStream.codecpar->channels > 1) {
+            kLogger.warning()
+                    << "Unknown channel layout -> using default layout"
+                    << channel_layout
+                    << "for"
+                    << avStream.codecpar->channels
+                    << "channels";
+        }
     }
     return channel_layout;
 }
@@ -459,7 +466,7 @@ QStringList SoundSourceProviderFFmpeg::getSupportedFileTypes() const {
     }
 
     if (!disabledInputFormats.isEmpty()) {
-        kLogger.info().noquote()
+        kLogger.debug().noquote()
                 << "Disabling untested input formats:"
                 << disabledInputFormats.join(QStringLiteral(", "));
     }
@@ -481,19 +488,25 @@ QString SoundSourceProviderFFmpeg::getVersionString() const {
     return QString::fromUtf8(av_version_info());
 }
 
-SoundSourceFFmpeg::SoundSourceFFmpeg(const QUrl& url)
+SoundSourceFFmpeg::SoundSourceFFmpeg(const QUrl& url, int wantedStreamIndex)
         : SoundSource(url),
           m_pavStream(nullptr),
           m_pavDecodedFrame(nullptr),
           m_seekPrerollFrameCount(0),
           m_pavPacket(av_packet_alloc()),
           m_pavResampledFrame(nullptr),
-          m_avutilVersion(avutil_version()) {
+          m_avutilVersion(avutil_version()),
+          m_wantedStreamIndex(wantedStreamIndex),
+          m_isLibfdk_aac(false) {
     DEBUG_ASSERT(m_pavPacket);
 #if LIBAVUTIL_VERSION_INT >= AV_VERSION_INT(57, 28, 100) // FFmpeg 5.1
     av_channel_layout_default(&m_avStreamChannelLayout, 0);
     av_channel_layout_default(&m_avResampledChannelLayout, 0);
 #endif
+}
+
+SoundSourceFFmpeg::SoundSourceFFmpeg(const QUrl& url)
+        : SoundSourceFFmpeg(url, kFirstAudioStream) {
 }
 
 SoundSourceFFmpeg::~SoundSourceFFmpeg() {
@@ -525,7 +538,11 @@ SoundSource::OpenResult SoundSourceFFmpeg::tryOpen(
     kLogger.debug()
             << "AVFormatContext"
             << "{ nb_streams" << m_pavInputFormatContext->nb_streams
-            << "| start_time" << m_pavInputFormatContext->start_time
+            << "| start_time"
+            << (m_pavInputFormatContext->start_time == AV_NOPTS_VALUE
+                               ? "AV_NOPTS_VALUE"
+                               : QString::number(
+                                         m_pavInputFormatContext->start_time))
             << "| duration" << m_pavInputFormatContext->duration
             << "| bit_rate" << m_pavInputFormatContext->bit_rate
             << "| packet_size" << m_pavInputFormatContext->packet_size
@@ -548,22 +565,35 @@ SoundSource::OpenResult SoundSourceFFmpeg::tryOpen(
     // Find the best stream
 #if LIBAVCODEC_VERSION_INT >= AV_VERSION_INT(59, 0, 100) // FFmpeg 5.0
     const AVCodec* pDecoder = nullptr;
-    const AVCodec* pAacDecoder = nullptr;
+    const AVCodec* pFdkAacDecoder = nullptr;
 #else
     // https://github.com/FFmpeg/FFmpeg/blob/dd17c86aa11feae2b86de054dd0679cc5f88ebab/doc/APIchanges#L175
     AVCodec* pDecoder = nullptr;
-    AVCodec* pAacDecoder = nullptr;
+    AVCodec* pFdkAacDecoder = nullptr;
 #endif
     const int av_find_best_stream_result = av_find_best_stream(
             m_pavInputFormatContext,
             AVMEDIA_TYPE_AUDIO,
-            /*wanted_stream_nb*/ -1,
+            m_wantedStreamIndex,
             /*related_stream*/ -1,
             &pDecoder,
             /*flags*/ 0);
     if (av_find_best_stream_result < 0) {
         switch (av_find_best_stream_result) {
         case AVERROR_STREAM_NOT_FOUND:
+            if (m_wantedStreamIndex >= 0) {
+                // This happens if m_wantedStreamIndex is not an audio stream
+                if (m_pavInputFormatContext->nb_streams <=
+                        static_cast<unsigned int>(m_wantedStreamIndex)) {
+                    kLogger.warning().noquote()
+                            << "cannot find stream" << m_wantedStreamIndex;
+                } else {
+                    kLogger.warning().noquote()
+                            << "stream" << m_wantedStreamIndex << "isn't an audio stream";
+                }
+                return OpenResult::Failed;
+            }
+            // called with kFirstAudioStream
             kLogger.warning()
                     << "av_find_best_stream() failed to find an audio stream";
             break;
@@ -582,22 +612,18 @@ SoundSource::OpenResult SoundSourceFFmpeg::tryOpen(
 
     if (pDecoder->id == AV_CODEC_ID_AAC ||
             pDecoder->id == AV_CODEC_ID_AAC_LATM) {
-        // We only allow AAC decoders that pass our seeking tests
-        if (std::strcmp(pDecoder->name, "aac") != 0 && std::strcmp(pDecoder->name, "aac_at") != 0) {
-            pAacDecoder = avcodec_find_decoder_by_name("aac");
-            if (pAacDecoder) {
-                pDecoder = pAacDecoder;
-            } else {
-                kLogger.warning()
-                        << "Internal aac decoder not found in your FFmpeg "
-                           "build."
-                        << "To enable AAC support, please install an FFmpeg "
-                           "version with the internal aac decoder enabled."
-                           "Note 1: The libfdk_aac decoder is no working properly "
-                           "with Mixxx, FFmpeg's internal AAC decoder does."
-                        << "Note 2: AAC decoding may be subject to patent "
-                           "restrictions, depending on your country.";
+        // Prefer Fraunhofer FDK AAC over internal AAC
+        // https://trac.ffmpeg.org/wiki/Encode/AAC
+        if (std::strcmp(pDecoder->name, "aac") == 0) {
+            pFdkAacDecoder = avcodec_find_decoder_by_name("libfdk_aac");
+            if (pFdkAacDecoder) {
+                pDecoder = pFdkAacDecoder;
             }
+        }
+
+        if (std::strcmp(pDecoder->name, "libfdk_aac") == 0) {
+            // Fraunhofer FDK AAC has an issue with flushing memory in the lead-in
+            m_isLibfdk_aac = true;
         }
     }
 
@@ -624,6 +650,9 @@ SoundSource::OpenResult SoundSourceFFmpeg::tryOpen(
                 << formatErrorString(avcodec_parameters_to_context_result);
         return SoundSource::OpenResult::Aborted;
     }
+
+    // Copy time base for random seeks:
+    pavCodecContext->pkt_timebase = pavStream->time_base;
 
     // Request output format
     pavCodecContext->request_sample_fmt = s_avSampleFormat;
@@ -981,12 +1010,6 @@ bool SoundSourceFFmpeg::adjustCurrentPosition(SINT startIndex) {
     // At the beginning of the stream, this is a negative position.
     auto seekIndex = startIndex - m_seekPrerollFrameCount;
 
-    // Seek to codec frame boundaries if the frame size is fixed and known
-    if (m_pavStream->codecpar->frame_size > 0) {
-        seekIndex -= seekIndex % m_pavCodecContext->frame_size;
-    }
-    DEBUG_ASSERT(seekIndex <= startIndex);
-
     if (m_frameBuffer.tryContinueReadingFrom(seekIndex)) {
         // No need to perform a costly seek operation. Just skip some buffered
         // samples and continue decoding at the current position.
@@ -994,7 +1017,20 @@ bool SoundSourceFFmpeg::adjustCurrentPosition(SINT startIndex) {
     }
 
     // Flush internal decoder state before seeking
-    avcodec_flush_buffers(m_pavCodecContext);
+    if (!m_isLibfdk_aac || seekIndex >= 0) {
+        // Fast: 0.6 us (Core Ultra 5 125U)
+        avcodec_flush_buffers(m_pavCodecContext);
+    } else {
+        // In case of libfdk_aac, we can't seek far enough into the lead in
+        // (to -m_seekPrerollFrameCount) to have a settled filter from silence.
+        // In the test SoundSourceProxyTest.seekBoundaries and  FFmpeg 4.4.2 it
+        // was limited to -661 instead of -2111. The workaround here is to reopen
+        // the codec which initializes all buffers with zero.
+        // Slow: 43 us (Core Ultra 5 125U)
+        const AVCodec* pCodec = m_pavCodecContext->codec;
+        avcodec_close(m_pavCodecContext);
+        avcodec_open2(m_pavCodecContext, pCodec, nullptr);
+    }
 
     // Seek to new position
     const int64_t seekTimestamp =
@@ -1075,48 +1111,47 @@ bool SoundSourceFFmpeg::consumeNextAVPacket(
     return true;
 }
 
-const CSAMPLE* SoundSourceFFmpeg::resampleDecodedAVFrame() {
+const CSAMPLE* SoundSourceFFmpeg::resampleDecodedAVFrame(AVFrame* pavDecodedFrame) {
     if (m_pSwrContext) {
         // Decoded frame must be resampled before reading
+        av_frame_unref(m_pavResampledFrame);
         m_pavResampledFrame->sample_rate = getSignalInfo().getSampleRate();
         m_pavResampledFrame->format = s_avSampleFormat;
 #if LIBAVUTIL_VERSION_INT >= AV_VERSION_INT(57, 28, 100) // FFmpeg 5.1
         av_channel_layout_copy(&m_pavResampledFrame->ch_layout, &m_avResampledChannelLayout);
-        if (m_pavDecodedFrame->ch_layout.order == AV_CHANNEL_ORDER_UNSPEC) {
+        if (pavDecodedFrame->ch_layout.order == AV_CHANNEL_ORDER_UNSPEC) {
             // Sometimes the channel layout is undefined.
-            av_channel_layout_copy(&m_pavDecodedFrame->ch_layout, &m_avStreamChannelLayout);
+            av_channel_layout_copy(&pavDecodedFrame->ch_layout, &m_avStreamChannelLayout);
         }
 #else
         m_pavResampledFrame->channel_layout = m_avResampledChannelLayout;
-        if (m_pavDecodedFrame->channel_layout == kavChannelLayoutUndefined) {
+        if (pavDecodedFrame->channel_layout == kavChannelLayoutUndefined) {
             // Sometimes the channel layout is undefined.
-            m_pavDecodedFrame->channel_layout = m_avStreamChannelLayout;
+            pavDecodedFrame->channel_layout = m_avStreamChannelLayout;
         }
 #endif
 #if VERBOSE_DEBUG_LOG
-        avTrace("Resampling decoded frame", *m_pavDecodedFrame);
+        avTrace("Resampling decoded frame", *pavDecodedFrame);
 #endif
         const auto swr_convert_frame_result = swr_convert_frame(
-                m_pSwrContext, m_pavResampledFrame, m_pavDecodedFrame);
+                m_pSwrContext, m_pavResampledFrame, pavDecodedFrame);
         if (swr_convert_frame_result != 0) {
             kLogger.warning().noquote()
                     << "swr_convert_frame() failed:"
                     << formatErrorString(swr_convert_frame_result);
-            // Discard decoded frame and abort after unrecoverable error
-            av_frame_unref(m_pavDecodedFrame);
             return nullptr;
         }
 #if VERBOSE_DEBUG_LOG
         avTrace("Received resampled frame", *m_pavResampledFrame);
 #endif
         DEBUG_ASSERT(m_pavResampledFrame->pts == AV_NOPTS_VALUE ||
-                m_pavResampledFrame->pts == m_pavDecodedFrame->pts);
-        DEBUG_ASSERT(m_pavResampledFrame->nb_samples == m_pavDecodedFrame->nb_samples);
+                m_pavResampledFrame->pts == pavDecodedFrame->pts);
+        DEBUG_ASSERT(m_pavResampledFrame->nb_samples == pavDecodedFrame->nb_samples);
         return reinterpret_cast<const CSAMPLE*>(
                 m_pavResampledFrame->extended_data[0]);
     } else {
         return reinterpret_cast<const CSAMPLE*>(
-                m_pavDecodedFrame->extended_data[0]);
+                pavDecodedFrame->extended_data[0]);
     }
 }
 
@@ -1196,7 +1231,6 @@ ReadableSampleFrames SoundSourceFFmpeg::readSampleFramesClamped(
                         (m_pavDecodedFrame->flags &
                                 (AV_FRAME_FLAG_CORRUPT |
                                         AV_FRAME_FLAG_DISCARD)) == 0) {
-                    av_frame_unref(m_pavDecodedFrame);
                     continue;
                 }
                 const auto decodedFrameCount = m_pavDecodedFrame->nb_samples;
@@ -1205,21 +1239,15 @@ ReadableSampleFrames SoundSourceFFmpeg::readSampleFramesClamped(
                         convertStreamTimeToFrameIndex(
                                 *m_pavStream, m_pavDecodedFrame->pts);
 
-                if (m_avutilVersion >= AV_VERSION_INT(56, 52, 100)) {
-                    // From ffmpeg 4.4 only audible samples are counted, i.e. any inaudible aka
-                    // "priming" samples are not included in nb_samples!
-                    // https://github.com/mixxxdj/mixxx/issues/10464
-                    if (streamFrameIndex < 0) {
 #if VERBOSE_DEBUG_LOG
-                        const auto inaudibleFrameCountUntilStartOfStream = -streamFrameIndex;
-                        kLogger.debug()
-                                << "Skipping"
-                                << inaudibleFrameCountUntilStartOfStream
-                                << "inaudible sample frames before the start of the stream";
-#endif
-                        streamFrameIndex = 0;
-                    }
+                if (streamFrameIndex < 0) {
+                    const SINT inaudibleFrameCountUntilStartOfStream = -streamFrameIndex;
+                    kLogger.info()
+                            << "Stream has"
+                            << inaudibleFrameCountUntilStartOfStream
+                            << "inaudible sample frames before the start";
                 }
+#endif
 
                 decodedFrameRange = IndexRange::forward(
                         streamFrameIndex,
@@ -1286,51 +1314,18 @@ ReadableSampleFrames SoundSourceFFmpeg::readSampleFramesClamped(
                     << "decodedFrameRange" << decodedFrameRange;
 #endif
 
-            const CSAMPLE* pDecodedSampleData = resampleDecodedAVFrame();
-            if (!pDecodedSampleData) {
+            const CSAMPLE* pDecodedSamples = resampleDecodedAVFrame(m_pavDecodedFrame);
+            if (!pDecodedSamples) {
                 // Invalidate current position and abort reading after unrecoverable error
                 m_frameBuffer.invalidate();
-                // Housekeeping before aborting to avoid memory leaks
-                av_frame_unref(m_pavDecodedFrame);
                 break;
             }
 
-            // The decoder may provide some lead-in and lead-out frames
-            // before the start position and after the end of the stream.
-            // Those frames need to be cut-off before consumption.
-            if (decodedFrameRange.start() < frameIndexRange().start()) {
-                const auto leadinRange = IndexRange::between(
-                        decodedFrameRange.start(),
-                        math_min(frameIndexRange().start(), decodedFrameRange.end()));
-                DEBUG_ASSERT(leadinRange.orientation() != IndexRange::Orientation::Backward);
-                if (leadinRange.orientation() == IndexRange::Orientation::Forward) {
 #if VERBOSE_DEBUG_LOG
-                    kLogger.debug()
-                            << "Cutting off lead-in"
-                            << leadinRange
-                            << "before"
-                            << frameIndexRange();
-#endif
-                    pDecodedSampleData += getSignalInfo().frames2samples(leadinRange.length());
-                    decodedFrameRange.shrinkFront(leadinRange.length());
-                }
+            if (!decodedFrameRange.empty()) {
+                kLogger.debug() << "First decoded sample value:" << pDecodedSamples[0];
             }
-            if (decodedFrameRange.end() > frameIndexRange().end()) {
-                const auto leadoutRange = IndexRange::between(
-                        math_max(frameIndexRange().end(), decodedFrameRange.start()),
-                        decodedFrameRange.end());
-                DEBUG_ASSERT(leadoutRange.orientation() != IndexRange::Orientation::Backward);
-                if (leadoutRange.orientation() == IndexRange::Orientation::Forward) {
-#if VERBOSE_DEBUG_LOG
-                    kLogger.debug()
-                            << "Cutting off lead-out"
-                            << leadoutRange
-                            << "beyond"
-                            << frameIndexRange();
 #endif
-                    decodedFrameRange.shrinkBack(leadoutRange.length());
-                }
-            }
 
 #if VERBOSE_DEBUG_LOG
             kLogger.debug()
@@ -1343,7 +1338,7 @@ ReadableSampleFrames SoundSourceFFmpeg::readSampleFramesClamped(
             const auto decodedSampleFrames = ReadableSampleFrames(
                     decodedFrameRange,
                     SampleBuffer::ReadableSlice(
-                            pDecodedSampleData,
+                            pDecodedSamples,
                             getSignalInfo().frames2samples(decodedFrameRange.length())));
             auto outputSampleFrames = WritableSampleFrames(
                     writableFrameRange,
@@ -1363,11 +1358,6 @@ ReadableSampleFrames SoundSourceFFmpeg::readSampleFramesClamped(
                     << "m_frameBuffer.bufferedRange()" << m_frameBuffer.bufferedRange()
                     << "writableFrameRange" << writableFrameRange;
 #endif
-
-            // Housekeeping before next decoding iteration
-            av_frame_unref(m_pavDecodedFrame);
-            av_frame_unref(m_pavResampledFrame);
-
             // The first loop condition (see below) should always be true
             // and has only been added to prevent infinite looping in case
             // of unexpected result values.